PairedLinkedLists

Documentation for PairedLinkedLists.

PairedLinkedLists.PairedLinkedLists

PairedLinkedLists.DoublyLinkedList

PairedLinkedLists.ListNode

PairedLinkedLists.PairedLinkedList

PairedLinkedLists.PairedListNode

PairedLinkedLists.TargetedLinkedList

PairedLinkedLists.TargetedListNode

PairedLinkedLists.addpartner!

PairedLinkedLists.deletenode!

PairedLinkedLists.getnode

PairedLinkedLists.haspartner

PairedLinkedLists.insertnode!

PairedLinkedLists.newnode

PairedLinkedLists.nodetype

PairedLinkedLists.removepartner!

PairedLinkedLists.PairedLinkedLists — Module

PairedLinkedLists.jl provides doubly-linked lists as well as "paired lists" that also contain links between nodes in two separate lists.

Array-like functionality is supported for DoublyLinkedList and PairedLinkedList:

    julia> using PairedLinkedLists

    julia> l = DoublyLinkedList{Int}(0)
    DoublyLinkedList{Int64}(0)

    julia> push!(l, 1:10...)
    DoublyLinkedList{Int64}(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10)

    julia> l[5] = 3;

    julia> popfirst!(l)
    0

    julia> l
    DoublyLinkedList{Int64}(1, 2, 3, 4, 3, 6, 7, 8, 9)

Supported methods include push!, pushfirst!, pop!, popfirst!, popat!, insert!, and splice!.

PairedLinkedLists.DoublyLinkedList — Type

l = DoublyLinkedList{::Type}()
l = DoublyLinkedList(elts...)

Create a DoublyLinkedList with with nodes containing data of a specified type.

The list contains its length len, a "dummy" node head at the beginning of the list, and a "dummy" node tail at the end of the list .

The first "real" node of a list l can be accessed with l.head.next. Similarly, the last "real" node can be accessed with l.tail.prev.

PairedLinkedLists.ListNode — Type

node = ListNode(list::DoublyLinkedList, data)

Create a ListNode belonging to the specified list. The node contains a reference list to the parent list, the provided data, a link prev to the preceding node, and a link next to the following node.

PairedLinkedLists.PairedLinkedList — Type

l = PairedLinkedList{::Type}()
l = PairedLinkedList{::Type}(elts...)

Create a PairedLinkedList with nodes containing data of a specified type.

The list contains its length len, a "dummy" node head at the beginning of the list, and a "dummy" node tail at the end of the list. The list also contains a reference to its "partner" list.

The first "real" node of a list l can be accessed with l.head.next. Similarly, the last "real" node can be accessed with l.tail.prev.

PairedLinkedLists.PairedListNode — Type

node = PairedListNode(list::PairedLinkedList, data)

Create a PairedListNode belonging to the specified list. The node contains a reference list to the parent list, the provided data, a link prev to the preceding node, a link next to the following node, and double-link partner to another PairedListNode.

A node's partner should always either be a reference to itself (denoting unpaired node) or a node belonging to the partner of its parent list.

PairedLinkedLists.TargetedLinkedList — Type

PairedLinkedLists.TargetedListNode — Type

node = TargetListNode(list::AbstractTargetLinkedList, data, [target::AbstractListNode])

Create a TargetListNode belonging to the specified list. The node contains a reference list to the parent list, the provided data, a link prev to the preceding node, a link next to the following node, and link partner to another list node.

A node's partner may be undefined or a node belonging to the partner of its parent list.

PairedLinkedLists.addpartner! — Method

addpartner!(node, partner_node)
addpartner!(list, partner_list)

Add a link between a the provided node or list and another object of the same type to be assigned its partner.

If the first object is a `PairedListNode' or a 'PairedLinkedList' and either object previously had a partner, the prior link is removed.

If the first object is a TargetedListNode or a TargetedLinkedList, the second object remains unchanged.

PairedLinkedLists.deletenode! — Method

deletenode!(node::ListNode)

Remove node from the list to which it belongs, update the list's length, and return the node.

PairedLinkedLists.getnode — Method

node = getnode(l::AbstractLinkedList, index)

Return the node at the specified index of the list.

PairedLinkedLists.haspartner — Method

haspartner(node::PairedListNode) -> Bool
haspartner(list::PairedListNode) -> Bool

Return true if the provided node or list has a partner, and false otherwise.

PairedLinkedLists.insertnode! — Method

insertnode!(node, prev)`

Insert node into a list after the preceding node prev, update the list's length, and return the node.

node and prev must belong to the same list.

PairedLinkedLists.newnode — Method

node = newnode(list, data)

Create an list node containing data of the appropriate type for the provided list. (e.g. a ListNode is created for a DoublyLinkedList).

PairedLinkedLists.nodetype — Method

t = nodetype(::AbstractLinkedList)
t = nodetype(::Type{<:AbstractLinkedList})

Return the type of the nodes contained in the list.

PairedLinkedLists.removepartner! — Method

removepartner!(node)

Remove the link between the node or list and its partner (if the object is already paired) and return node.

If the object is a PairedListNode or PairedLinkedList, the link will be deleted from both the object and its partner.

If the object is a TargetedListNode or PairedLinkedList, the link will be deleted from only the object.