PairedLinkedLists

Documentation for PairedLinkedLists.

PairedLinkedLists.PairedLinkedLists

PairedLinkedLists.DoublyLinkedList

PairedLinkedLists.ListDataIterator

PairedLinkedLists.ListDataIterator

PairedLinkedLists.ListNode

PairedLinkedLists.ListNodeIterator

PairedLinkedLists.ListNodeIterator

PairedLinkedLists.PairedLinkedList

PairedLinkedLists.PairedListNode

PairedLinkedLists.PairedSkipList

PairedLinkedLists.PairedSkipNode

PairedLinkedLists.SkipList

PairedLinkedLists.SkipNode

PairedLinkedLists.TargetedLinkedList

PairedLinkedLists.TargetedListNode

PairedLinkedLists.addtarget!

PairedLinkedLists.athead

PairedLinkedLists.attail

PairedLinkedLists.deletenode!

PairedLinkedLists.deletenode!

PairedLinkedLists.getnode

PairedLinkedLists.hastarget

PairedLinkedLists.head

PairedLinkedLists.insertafter!

PairedLinkedLists.insertbefore!

PairedLinkedLists.listtype

PairedLinkedLists.newnode

PairedLinkedLists.nodetype

PairedLinkedLists.removetarget!

PairedLinkedLists.tail

PairedLinkedLists.PairedLinkedLists — Module

PairedLinkedLists.jl provides implementations for doubly-linked lists and skip lists, as well as "paired lists" that also contain inter-list links between nodes in two separate lists.

See also DoublyLinkedList, PairedLinkedList, SkipList, PairedSkipList

PairedLinkedLists.DoublyLinkedList — Type

l = DoublyLinkedList{::Type}()
l = DoublyLinkedList(elts...)

Create a DoublyLinkedList made up of ListNodes with with nodes containing data of a specified type.

The list contains its length len, a "dummy" node head at the beginning of the list, and a "dummy" node tail at the end of the list .

The first "real" node of a list l can be accessed with l.head.next or head(l). Similarly, the last "real" node can be accessed with l.tail.prev or tail(l).

See also ListNode, SkipList, PairedLinkedList, TargetedLinkedList

PairedLinkedLists.ListDataIterator — Type

ListDataIterator(list; rev=false)

Returns an iterator over the data contained in a linked list, starting at the specified node start.

If rev is true, the iterator will advance toward the head of the list. Otherwise, it will advance toward the tail of the list.

PairedLinkedLists.ListDataIterator — Method

ListDataIterator(list; rev=false)

Returns an iterator over the data contained in a linked list.

If rev is true, the iterator will start at the tail of the list and advance toward the head. Otherwise, it will start at the head of the list and advance toward the tail.

PairedLinkedLists.ListNode — Type

node = ListNode(list::DoublyLinkedList [, data])

Create a ListNode belonging to the specified list. The node contains a reference list to the parent list and contains the provided data, but it has no specific insertion point into list (see insertafter!).

node.prev and node.next represent the previous and next nodes, respectively, of a list.

See also DoublyLinkedList, PairedListNode, TargetedListNode.

PairedLinkedLists.ListNodeIterator — Type

ListNodeIterator(start; rev=false)

Returns an iterator over the nodes of a linked list, starting at the specified node start.

If rev is true, the iterator will advance toward the head of the list. Otherwise, it will advance toward the tail of the list.

PairedLinkedLists.ListNodeIterator — Method

ListNodeIterator(list; rev=false)

Returns an iterator over the nodes of a linked list.

If rev is true, the iterator will start at the tail of the list and advance toward the head. Otherwise, it will start at the head of the list and advance toward the tail.

PairedLinkedLists.PairedLinkedList — Type

l = PairedLinkedList{::Type}()
l = PairedLinkedList{::Type}(elts...)

Create a PairedLinkedList made up of PairListNodes containing data of a specified type. Each node can have an inter-list link to a node belonging to the list's target.

The list contains its length len, a "dummy" node head at the beginning of the list, and a "dummy" node tail at the end of the list .

The first "real" node of a list l can be accessed with l.head.next or head(l). Similarly, the last "real" node can be accessed with l.tail.prev or tail(l).

See also PairedListNode, PairedSkipList, DoublyLinkedList, TargetedLinkedList

PairedLinkedLists.PairedListNode — Type

node = PairedListNode(list::PairedLinkedList, data)

Create a PairedListNode belonging to the specified list. The node contains a reference list to the parent list and contains the provided data, but it has no specific insertion point into list (see insertafter!).

node.prev and node.next represent the previous and next nodes, respectively, of a list.

A node's target should always either be a reference to itself (denoting unpaired node) or a node belonging to the target of its parent list.

The target link is assumed to be reciprocated for a PairedListNode. For example, node === node.target.target should be true. For one-way inter-list links, see TargetedListNode.

See also PairedLinkedList, ListNode, TargetedListNode.

PairedLinkedLists.PairedSkipList — Type

l = PairedSkipList{::Type}(; sortedby=identity, skipfactor=2)
l = PairedSkipList{::Type}(elts...; sortedby=identity, skipfactor=2)

Create a PairedSkipList with nodes containing data of a specified type. Each node can have an inter-list link to a node belonging to the list's target.

The list contains its length len, a "dummy" node head at the beginning of the list, and a "dummy" node tail at the end of the list. The list also contains a reference to its "target" list.

The first "real" node of a list l can be accessed with l.head.next. Similarly, the last "real" node can be accessed with l.tail.prev.

The skipfactor of the list describes the average number of nodes "skipped" by the above level.

The ordering of the list can be specified by a function, sortedby.

See also SkipList, PairedSkipNode

PairedLinkedLists.PairedSkipNode — Type

node = PairedSkipNode(list::SkipList [, data])

Create a PairedSkipNode belonging to the specified list. The node contains a reference list to the parent skip list and contains the provided data, but it has no specific insertion point into list (see insertafter!).

node.prev and node.next represent the previous and next nodes, respectively, of a list. node.up and node.bottom represent the nodes in adjacent levels within the skip list data structure.

A node's target should always either be a reference to itself (denoting unpaired node) or a node belonging to the target of its parent list.

The target link is assumed to be reciprocated for a PairedSkipNode. For example, node === node.target.target should be true.

See also PairedSkipList, SkipNode

PairedLinkedLists.SkipList — Type

l = SkipList{::Type}(; sortedby=identity, skipfactor=2)
l = SkipList{::Type}(elts...; sortedby=identity, skipfactor=2)

Create a SkipList made of up SkipNodes containing data of a specified type.

The list contains a "dummy" node head at the beginning of the list and a "dummy" node tail at the end of the list.

The node at the head of the topmost level of the datastructure can be accessed by l.top.

The first "real" node of a list l can be accessed with l.head.next or head(l). Similarly, the last "real" node can be accessed with l.tail.prev or tail(l).

The skipfactor of the list describes the average number of nodes "skipped" by the above level.

The ordering of the list can be specified by a function, sortedby.

See also PairedSkipList, SkipNode

PairedLinkedLists.SkipNode — Type

node = SkipNode(list::SkipList [, data])

Create a SkipNode belonging to the specified list. The node contains a reference list to the parent skip list and contains the provided data, but it has no specific insertion point into list (see insertafter!).

node.prev and node.next represent the previous and next nodes, respectively, of a list. node.up and node.bottom represent the nodes in adjacent levels within the skip list data structure.

See also SkipList, PairedSkipNode

PairedLinkedLists.TargetedLinkedList — Type

l = TargetLinkedList{T,R}()
l = TargetLinkedList{T,R}(elts...)
l = TargetLinkedList(list)

Create a TargetLinkedList made up of TargetedListNodes containing data of a specified type. Each node can have an inter-list link to a node belonging to the list's target.

The list contains its length len, a "dummy" node head at the beginning of the list, and a "dummy" node tail at the end of the list. The list also contains a reference to its "target" list.

The first "real" node of a list l can be accessed with l.head.next or head(l). Similarly, the last "real" node can be accessed with l.tail.prev or tail(l).

See also TargetedListNode, DoublyLinkedList, PairedLinkedList

PairedLinkedLists.TargetedListNode — Type

node = TargetListNode(list::AbstractTargetLinkedList, data, [target::AbstractListNode])

Create a TargetedListNode belonging to the specified list. The node contains a reference list to the parent list and contains the provided data, but it has no specific insertion point into list (see insertafter!).

node.prev and node.next represent the previous and next nodes, respectively, of a list.

A node's target should always either be a reference to itself (denoting unpaired node) or a node belonging to the target of its parent list.

The target link is not assumed to be reciprocated for a PairedListNode, as the targeted node may not even have a target field. For guranteed two-way inter-list links, see PairedListNode.

See also TargetedLinkedList, PairedListNode, ListNode.

PairedLinkedLists.addtarget! — Method

addtarget!(node, target_node)
addtarget!(list, target_list)

Add a link between a the provided node or list and another object of the same type to be assigned its target.

If the first object is a `PairedListNode' or a 'PairedLinkedList' and either object previously had a target, the prior link is removed.

If the first object is a TargetedListNode or a TargetedLinkedList, the second object remains unchanged.

See also hastarget, removetarget!

PairedLinkedLists.athead — Method

athead(node) -> Bool

Return true if the node is the "dummy" node at the beginning of the list, and false otherwise.

PairedLinkedLists.attail — Method

attail(node) -> Bool

Return true if the node is the "dummy" node at the end of the list, and false otherwise.

PairedLinkedLists.deletenode! — Method

deletenode!(node::SkipNode)

Remove node from the list to which it belongs, update the list's length, and return the node.

The node can be at any level of the skip list, and all nodes directly above or below will also be removed.

PairedLinkedLists.deletenode! — Method

deletenode!(node::ListNode)

Remove node from the list to which it belongs, update the list's length, and return the node.

PairedLinkedLists.getnode — Method

node = getnode(l::AbstractList, index)

Return the node at the specified index of the list.

PairedLinkedLists.hastarget — Method

hastarget(node) -> Bool
hastarget(list) -> Bool

Return true if the provided node or list has a target, and false otherwise.

See also addtarget!, removetarget!

PairedLinkedLists.head — Method

node = head(list)

Returns the first "real" node in the list. Note that this is not the same as list.head, which is a "dummy" node.

PairedLinkedLists.insertafter! — Method

insertafter!(node, prev)`

Insert node into a list after the preceding node prev, update the list's length, and return the node.

node and prev must belong to the same list.

PairedLinkedLists.insertbefore! — Method

insertbefore!(node, next)`

Insert node into a list before the subsequent node next, update the list's length, and return the node.

node and next must belong to the same list.

PairedLinkedLists.listtype — Method

t = listtype(::AbstractNode)
t = listtype(::Type{<:AbstractNode})

Return the type of list that can contain the provided type of node.

PairedLinkedLists.newnode — Method

node = newnode(list, data)

Create an list node containing data of the appropriate type for the provided list. (e.g. a ListNode is created for a DoublyLinkedList).

The node is disconnected from the list (see insertafter!).

PairedLinkedLists.nodetype — Method

t = nodetype(::AbstractList)
t = nodetype(::Type{<:AbstractList})

Return the type of the nodes contained in the list.

PairedLinkedLists.removetarget! — Method

removetarget!(node)

Remove the link between the node or list and its target (if the object is already paired) and return node.

If the object is a PairedListNode or PairedLinkedList, the link will be deleted from both the object and its target.

If the object is a TargetedListNode or PairedLinkedList, the link will be deleted from only the object.

See also hastarget, addtarget!

PairedLinkedLists.tail — Method

node = head(list)

Returns the last "real" node in the list. Note that this is not the same as list.tail, which is a "dummy" node.